# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for assigning reviewers to pull requests.",
        "title": "PR Reviewer Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/users/getReview": {
            "get": {
                "description": "Возвращает PR, где пользователь выступает ревьювером.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Получить PR пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор пользователя",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserReviewResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/users/setIsActive": {
            "post": {
                "description": "Ставит или снимает флаг активности пользователя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Обновить активность пользователя",
                "parameters": [
                    {
                        "description": "Параметры активности",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ErrorBody": {
            "description": "Содержит код и сообщение ошибки.",
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "description": "Код ошибки.",
                    "type": "string",
                    "example": "NOT_FOUND"
                },
                "message": {
                    "description": "Сообщение ошибки.",
                    "type": "string",
                    "example": "resource not found"
                }
            }
        },
        "ErrorResponse": {
            "description": "Стандартный ответ с ошибкой.",
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "description": "Детали ошибки.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ErrorBody"
                        }
                    ]
                }
            }
        },
...
        "PullRequestShort": {
            "description": "Короткое описание PR.",
            "type": "object",
            "required": [
                "author_id",
                "pull_request_id",
                "pull_request_name",
                "status"
            ],
            "properties": {
                "author_id": {
                    "description": "Автор PR.",
                    "type": "string",
                    "example": "u1"
                },
                "pull_request_id": {
                    "description": "Идентификатор PR.",
                    "type": "string",
                    "example": "pr-1001"
                },
                "pull_request_name": {
                    "description": "Название PR.",
                    "type": "string",
                    "example": "Add search endpoint"
                },
                "status": {
                    "description": "Статус PR.",
                    "type": "string",
                    "example": "OPEN"
                }
            }
        },
...
        "User": {
            "description": "Представление пользователя.",
            "type": "object",
            "required": [
                "is_active",
                "team_name",
                "user_id",
                "username"
            ],
            "properties": {
                "is_active": {
                    "description": "Флаг активности.",
                    "type": "boolean",
                    "example": true
                },
                "team_name": {
                    "description": "Название команды.",
                    "type": "string",
                    "example": "backend"
                },
                "user_id": {
                    "description": "Идентификатор пользователя.",
                    "type": "string",
                    "example": "u2"
                },
                "username": {
                    "description": "Имя пользователя.",
                    "type": "string",
                    "example": "Bob"
                }
            }
        },
        "UserRequest": {
            "description": "Запрос на смену активности пользователя.",
            "type": "object",
            "required": [
                "is_active",
                "user_id"
            ],
            "properties": {
                "is_active": {
                    "description": "Значение флага активности.",
                    "type": "boolean",
                    "example": false
                },
                "user_id": {
                    "description": "Идентификатор пользователя.",
                    "type": "string",
                    "example": "u2"
                }
            }
        },
        "UserResponse": {
            "description": "Ответ с пользователем.",
            "type": "object",
            "required": [
                "user"
            ],
            "properties": {
                "user": {
                    "description": "Пользователь.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/User"
                        }
                    ]
                }
            }
        },
        "UserReviewResponse": {
            "description": "PR, где пользователь ревьювер.",
            "type": "object",
            "required": [
                "pull_requests",
                "user_id"
            ],
            "properties": {
                "pull_requests": {
                    "description": "Список PR.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/PullRequestShort"
                    }
                },
                "user_id": {
                    "description": "Идентификатор пользователя.",
                    "type": "string",
                    "example": "u2"
                }
            }
        }
    }
}